use decorators for common gate extensions
