+++ qiskit/circuit/instructionset.py
@@ -23,6 +23,10 @@ class InstructionSet:
         """Return number of instructions in set"""
         return len(self.instructions)
 
+    def __getitem__(self, i):
+        """Return instruction at index"""
+        return self.instructions[i]
+
     def add(self, gate):
         """Add instruction to set."""
         if not isinstance(gate, Instruction):
