+++ qiskit/extensions/standard/cx.py
@@ -12,8 +12,8 @@ controlled-NOT gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _control_target_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.cxbase import CXBase
@@ -50,20 +50,9 @@ class CnotGate(Gate):
         self._modifiers(circ.cx(self.qargs[0], self.qargs[1]))
 
 
+@_control_target_gate
 def cx(self, ctl, tgt):
     """Apply CX from ctl to tgt."""
-    if isinstance(ctl, QuantumRegister):
-        ctl = [(ctl, i) for i in range(len(ctl))]
-    if isinstance(tgt, QuantumRegister):
-        tgt = [(tgt, i) for i in range(len(tgt))]
-
-    if ctl and tgt and isinstance(ctl, list) and \
-       isinstance(tgt, list) and len(ctl) == len(tgt):
-        instructions = InstructionSet()
-        for ictl, itgt in zip(ctl, tgt):
-            instructions.add(self.cx(ictl, itgt))
-        return instructions
-
     self._check_qubit(ctl)
     self._check_qubit(tgt)
     self._check_dups([ctl, tgt])
