+++ qiskit/extensions/standard/cxbase.py
@@ -10,8 +10,7 @@ Fundamental controlled-NOT gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
-from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _control_target_gate
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 
 
@@ -31,20 +30,9 @@ class CXBase(Gate):  # pylint: disable=abstract-method
         self._modifiers(circ.cx_base(self.qargs[0], self.qargs[1]))
 
 
+@_control_target_gate
 def cx_base(self, ctl, tgt):
     """Apply CX ctl, tgt."""
-    if isinstance(ctl, QuantumRegister):
-        ctl = [(ctl, i) for i in range(len(ctl))]
-    if isinstance(tgt, QuantumRegister):
-        tgt = [(tgt, i) for i in range(len(tgt))]
-
-    if ctl and tgt and isinstance(ctl, list) and \
-       isinstance(tgt, list) and len(ctl) == len(tgt):
-        instructions = InstructionSet()
-        for ictl, itgt in zip(ctl, tgt):
-            instructions.add(self.cx_base(ictl, itgt))
-        return instructions
-
     self._check_qubit(ctl)
     self._check_qubit(tgt)
     self._check_dups([ctl, tgt])
