+++ qiskit/extensions/standard/cy.py
@@ -12,8 +12,8 @@ controlled-Y gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _control_target_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.s import SGate
@@ -56,20 +56,9 @@ class CyGate(Gate):
         self._modifiers(circ.cy(self.qargs[0], self.qargs[1]))
 
 
+@_control_target_gate
 def cy(self, ctl, tgt):
     """Apply CY to circuit."""
-    if isinstance(ctl, QuantumRegister):
-        ctl = [(ctl, i) for i in range(len(ctl))]
-    if isinstance(tgt, QuantumRegister):
-        tgt = [(tgt, i) for i in range(len(tgt))]
-
-    if ctl and tgt and isinstance(ctl, list) and \
-       isinstance(tgt, list) and len(ctl) == len(tgt):
-        instructions = InstructionSet()
-        for ictl, itgt in zip(ctl, tgt):
-            instructions.add(self.cy(ictl, itgt))
-        return instructions
-
     self._check_qubit(ctl)
     self._check_qubit(tgt)
     self._check_dups([ctl, tgt])
