+++ qiskit/extensions/standard/h.py
@@ -12,8 +12,8 @@ Hadamard gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.qasm import pi
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
@@ -51,18 +51,9 @@ class HGate(Gate):
         self._modifiers(circ.h(self.qargs[0]))
 
 
+@_1q_gate
 def h(self, q):
     """Apply H to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.h(qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(HGate(q, self))
