+++ qiskit/extensions/standard/iden.py
@@ -12,8 +12,8 @@ Identity gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.ubase import UBase
@@ -47,18 +47,9 @@ class IdGate(Gate):
         self._modifiers(circ.iden(self.qargs[0]))
 
 
+@_1q_gate
 def iden(self, q):
     """Apply Identity to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.iden(qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(IdGate(q, self))
