+++ qiskit/extensions/standard/rx.py
@@ -12,8 +12,8 @@ Rotation around the x-axis.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.qasm import pi
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
@@ -56,18 +56,9 @@ class RXGate(Gate):
         self._modifiers(circ.rx(self.param[0], self.qargs[0]))
 
 
+@_1q_gate
 def rx(self, theta, q):
     """Apply Rx to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.rx(theta, qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(RXGate(theta, q, self))
