+++ qiskit/extensions/standard/rz.py
@@ -12,8 +12,8 @@ Rotation around the z-axis.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.u1 import U1Gate
@@ -55,18 +55,9 @@ class RZGate(Gate):
         self._modifiers(circ.rz(self.param[0], self.qargs[0]))
 
 
+@_1q_gate
 def rz(self, phi, q):
     """Apply Rz to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.rz(phi, qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(RZGate(phi, q, self))
