+++ qiskit/extensions/standard/rzz.py
@@ -10,8 +10,8 @@ two-qubit ZZ-rotation gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _control_target_gate
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard.u1 import U1Gate
@@ -54,20 +54,9 @@ class RZZGate(Gate):
         self._modifiers(circ.rzz(self.param[0], self.qargs[0], self.qargs[1]))
 
 
+@_control_target_gate
 def rzz(self, theta, ctl, tgt):
     """Apply RZZ to circuit."""
-    if isinstance(ctl, QuantumRegister):
-        ctl = [(ctl, i) for i in range(len(ctl))]
-    if isinstance(tgt, QuantumRegister):
-        tgt = [(tgt, i) for i in range(len(tgt))]
-
-    if ctl and tgt and isinstance(ctl, list) and \
-       isinstance(tgt, list) and len(ctl) == len(tgt):
-        instructions = InstructionSet()
-        for ictl, itgt in zip(ctl, tgt):
-            instructions.add(self.rzz(theta, ictl, itgt))
-        return instructions
-
     self._check_qubit(ctl)
     self._check_qubit(tgt)
     self._check_dups([ctl, tgt])
