+++ qiskit/extensions/standard/t.py
@@ -12,8 +12,8 @@ T=sqrt(S) phase gate or its inverse.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.qasm import pi
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
@@ -86,34 +86,16 @@ class TdgGate(Gate):
         return inv
 
 
+@_1q_gate
 def t(self, q):
     """Apply T to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.t(qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(TGate(q, self))
 
 
+@_1q_gate
 def tdg(self, q):
     """Apply Tdg to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.tdg(qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(TdgGate(q, self))
