+++ qiskit/extensions/standard/u0.py
@@ -12,8 +12,8 @@ Single qubit gate cycle idle.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.ubase import UBase
@@ -47,18 +47,9 @@ class U0Gate(Gate):
         self._modifiers(circ.u0(self.param[0], self.qargs[0]))
 
 
+@_1q_gate
 def u0(self, m, q):
     """Apply u0 with length m to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.u0(m, qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(U0Gate(m, q, self))
