+++ qiskit/extensions/standard/u3.py
@@ -12,8 +12,8 @@ Two-pulse single-qubit gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.ubase import UBase
@@ -56,18 +56,9 @@ class U3Gate(Gate):
                                 self.qargs[0]))
 
 
+@_1q_gate
 def u3(self, theta, phi, lam, q):
     """Apply u3 to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.u3(theta, phi, lam, qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(U3Gate(theta, phi, lam, q, self))
