+++ qiskit/extensions/standard/x.py
@@ -12,8 +12,8 @@ Pauli X (bit-flip) gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.qasm import pi
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
@@ -53,18 +53,9 @@ class XGate(Gate):
         self._modifiers(circ.x(self.qargs[0]))
 
 
+@_1q_gate
 def x(self, q):
     """Apply X to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.x(qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(XGate(q, self))
