+++ qiskit/extensions/standard/z.py
@@ -12,8 +12,8 @@ Pauli Z (phase-flip) gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.quantumcircuit import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.qasm import pi
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
@@ -48,18 +48,9 @@ class ZGate(Gate):
         self._modifiers(circ.z(self.qargs[0]))
 
 
+@_1q_gate
 def z(self, q):
     """Apply Z to q."""
-    if isinstance(q, QuantumRegister):
-        q = [(q, j) for j in range(len(q))]
-
-    if q and isinstance(q, list):
-        instructions = InstructionSet()
-        for qubit in q:
-            self._check_qubit(qubit)
-            instructions.add(self.z(qubit))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(ZGate(q, self))
